Update: Follow Service [중복 팔로우, 자기 자신 팔로우, 언팔로우 에러 핸들링 추가], ChatController [존재하지 않은 채팅방 삭제 에러 핸들링 추가]

+++ src/main/java/okky/team/chawchaw/follow/FollowServiceImpl.java
@@ -18,22 +18,27 @@ public class FollowServiceImpl implements FollowService{
 
     @Override
     @Transactional(readOnly = false)
-    public Long addFollow(UserEntity userFrom, Long userTo) {
+    public Boolean addFollow(UserEntity userFrom, Long userTo) {
 
         UserEntity user = userRepository.findById(userTo).orElseThrow(() -> new UsernameNotFoundException("not found user"));
 
-        FollowEntity result = followRepository.save(new FollowEntity(userFrom, user));
-
-        return result.getId();
+        if (!followRepository.isFollow(userFrom.getId(), userTo)) {
+            followRepository.save(new FollowEntity(userFrom, user));
+            return true;
+        }
+        return false;
     }
 
     @Override
     @Transactional(readOnly = false)
-    public void deleteFollow(UserEntity userFrom, Long userTo) {
+    public Boolean deleteFollow(UserEntity userFrom, Long userTo) {
 
         UserEntity user = userRepository.findById(userTo).orElseThrow(() -> new UsernameNotFoundException("not found user"));
 
-        followRepository.removeByUserFromAndUserTo(userFrom, user);
-
+        if (followRepository.isFollow(userFrom.getId(), userTo)) {
+            followRepository.removeByUserFromAndUserTo(userFrom, user);
+            return true;
+        }
+        return false;
     }
 }
